Extract logic to usecase
